pyproject.toml
[tool.poetry]
name = "unscheduler"
version = "0.1.0"
description = "Program to make and manage an unschedule"
authors = ["James J. Hamlin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.9.0.post0"
matplotlib = "^3.10.5"
pytest = "^8.4.1"
watchdog = "^6.0.0"
textual = "^5.3.0"
humanize = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
